tests/wpt/web-platform-tests/css/css-will-change/parsing/will-change-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Will Change Test: getComputedStyle().willChange</title>
<link rel="help" href="https://drafts.csswg.org/css-will-change/#propdef-will-change">
<meta name="assert" content="will-change computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("will-change", "auto");

// <animateable-feature> = scroll-position | contents | <custom-ident>
test_computed_value("will-change", "scroll-position");
test_computed_value("will-change", "contents");
test_computed_value("will-change", "transform");
test_computed_value("will-change", "background-color");

test_computed_value("will-change", "scroll-position, contents");
test_computed_value("will-change", "scroll-position, transform");
test_computed_value("will-change", "contents, transform");
test_computed_value("will-change", "transform, background-color");
</script>
</body>
</html>
